+++ src/test/scala/scalismo/registration/KernelTransformationTests.scala
@@ -36,7 +36,7 @@ class KernelTransformationTests extends ScalismoTestSuite {
   describe("The Nystroem approximation of a Kernel matrix") {
     it("Is close enough to a scalar valued kernel matrix") {
       val kernel = DiagonalKernel[_1D](GaussianKernel[_1D](20))
-      val domain = BoxDomain(-5.0f, 195.0f)
+      val domain = BoxDomain(-5.0, 195.0)
 
       val sampler = UniformSampler(domain, 500)
 
@@ -52,14 +52,14 @@ class KernelTransformationTests extends ScalismoTestSuite {
       for ((x, _) <- sampler.sample; (y, _) <- sampler.sample) {
         val v1 = kernel(x, y)(0, 0)
         val v2 = approxKernel(x, y)
-        v2.toFloat should be(v1 +- 0.01f)
+        v2 should be(v1 +- 0.01)
       }
     }
 
     it("Its eigenvalues are close enough to the real eigenvalues for 1D") {
       val kernelDim = 1
       val scalarKernel = DiagonalKernel[_1D](GaussianKernel[_1D](10))
-      val domain = BoxDomain(0.0f, 10.0f)
+      val domain = BoxDomain(0.0, 10.0)
       val numPoints = 500
       val sampler = UniformSampler(domain, numPoints)
       val (points, _) = sampler.sample.unzip
@@ -76,7 +76,7 @@ class KernelTransformationTests extends ScalismoTestSuite {
       val (_, realLambdas, _) = RandomSVD.computeSVD(realKernelMatrix * (1.0 / numPoints), eigPairsApprox.size)
 
       for (l <- approxLambdas.zipWithIndex)
-        l._1 should be(realLambdas(l._2).toFloat +- 0.1f)
+        l._1 should be(realLambdas(l._2) +- 0.1)
 
     }
 
@@ -85,7 +85,7 @@ class KernelTransformationTests extends ScalismoTestSuite {
       val kernelDim = 2
       val scalarKernel = GaussianKernel[_2D](10)
       val ndKernel = DiagonalKernel[_2D](scalarKernel)
-      val domain = BoxDomain((0.0f, 0.0f), (5.0f, 5.0f))
+      val domain = BoxDomain((0.0, 0.0), (5.0, 5.0))
       val sampler = UniformSampler(domain, 400)
       val (pts, _) = sampler.sample.unzip
 
@@ -100,13 +100,13 @@ class KernelTransformationTests extends ScalismoTestSuite {
 
       val (_, realLambdas, _) = RandomSVD.computeSVD(realKernelMatrix * (1.0 / pts.size), eigPairsApprox.size)
       for (l <- approxLambdas.zipWithIndex)
-        l._1 should be(realLambdas(l._2).toFloat +- 0.1)
+        l._1 should be(realLambdas(l._2) +- 0.1)
 
     }
 
     it("It leads to orthogonal basis functions on the domain (-5, 5)") {
       val kernel = DiagonalKernel[_1D](GaussianKernel[_1D](1.0))
-      val domain = BoxDomain(-5.0f, 5.0f)
+      val domain = BoxDomain(-5.0, 5.0)
       val grid = DiscreteImageDomain(domain.origin, domain.extent * (1.0 / 1000.0), IntVector(1000))
       val sampler = GridSampler(grid)
